// src/apis/models/menu-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mcafe V2 API
 * Mcafe V2 API 명세서
 *
 * The version of the OpenAPI document: 2.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface MenuDto
 */
export interface MenuDto {
  /**
   *
   * @type {string}
   * @memberof MenuDto
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof MenuDto
   */
  code: string;
  /**
   *
   * @type {number}
   * @memberof MenuDto
   */
  price: number;
  /**
   *
   * @type {string}
   * @memberof MenuDto
   */
  unit: string;
  /**
   *
   * @type {number}
   * @memberof MenuDto
   */
  stock: number;
}